feat: rename NFTListPage to NFTList
