+++ webapp/src/components/NFTList/ArrayFilter/ArrayFilter.types.ts
