+++ webapp/src/components/NFTList/FiltersMenu/FiltersMenu.types.ts
