+++ webapp/src/components/NFTList/TextFilter/TextFilter.types.ts
