--- webapp/src/components/NFTListPage/index.ts
@@ -1,2 +1,0 @@
-import NFTListPage from './NFTListPage.container'
-export { NFTListPage }
